Update mainwindow.cpp
added board and sample buttons feel free to move the stuff around a delete the buttons we can move then down or to the side using the grid layout object. I have 2 grid layout objects one for the overall layout and one for the actual board. I know there is a way to constrain the board to a section of the outer grid and I will try to do it later.  Like a said earlier the board should zoom with the background so I think we should leave the zoom functions there but we can disable the background so it isn't distracting though I wanted it there for demonstration purposes.  I uses a QsignalMapper to correctly detect the region of the board that is selected.